TaskQueue creation working
Decided to add in a property for a TaskQueue's associated
AlchemicalNetwork's scoped key, and use this as the basis for its own
gufe key. This allows us to rely on the same DB constraints that we do
for all GufeTokenizables in terms of ensuring uniqueness, but makes it
so that uniqueness of the TaskQueue's key is independent of its other
attributes, such as `weight`, which can change with time or be identical
to another TaskQueue in the same Scope.