Implemented taskqueue with head and tail nodes
Also added ability to set taskqueue weight.
Put a lot of effort into trying to make Cypher do most of the work to
take advantage of transactions where possible, but this is a challenge.
Still working through adding tasks to a queue and getting the
relationship connections right in a single transaction.