Finally identified and hopefully resolved a *very* subtle py2neo bug
If a node in neo4j gets created as id 0, then py2neo's Subgraph object
won't deduplicate that object when set operations include it. We rely on
this functionality for reconstituting gufe objects, and this bug would
rear its head seemingly at random in our test suite. In production it
is unlikely to be an issue for a neo4j instance that is long-lived, and
therefore long past creating an object with id 0.

I'll raise an issue on `py2neo`'s issue tracker, but that library is
looking fairly unmaintained these days, so we may be better off building
around the official neo4j python driver as we go forward. We'll limp
along with our approach for now.