A load of new `get_` methods on AlchemiscaleClient
These methods allow for grabbing objects associated with other objects,
such as the transformations in a network, the networks a transformation
is a part of, chemicalsystems associated with a transformation, etc.

All these methods deal in ScopedKeys, and so they are intended to be
very fast. Also, it allows retrieval of transformation ScopedKeys for
a given network, which can directly be used for result retrieval, even
if there is a mismatch in the gufe keys living in the state store and
those that would be given client-side for those transformations.