Added `get_network_results` method to `AlchemiscaleClient`
Closes #154.

There are a couple ways we could go about this. It's unclear which one
will yield higher performance. Will try both and compare:
1. retrieve results for each `Transformation` as a separate task for a
   process pool, with each using `asyncio` separately
2. toss all `ProtocolDAGResult` requests into the same `asyncio` event loop

(1) may make better use of available CPU resources, since a bottleneck
when retrieving many or large `ProtocolDAGResult`s is deserialization
from JSON. So that deserialization can be done over many processes, but
an additional pickle serialization/deserialization is required to return
these objects to the parent process.

(2) would be more efficient if the bottleneck really is just network
latency, not deserialization.

We may end up providing both (1) and (2), and make them switchable via a
kwarg. This will allow users to choose the approach that yields better
performance for them, which can be geography and local network
speed-dependent.