Make test env use latest release of openfe and gufe instead of `main`
Now that `openfe` and `gufe` are operating at regular release cadences,
I think it makes sense for us to perform testing against the latest
releases of these on `conda-forge` instead of testing against the `main`
branches of them. In particular, this saves us the need of having to
manually track their dependencies, such as the recently added
`kartograf` dependency in `openfe`.

I think this is generally a safe move, since we only ever deploy
production envs with releases of these packages too (pinned to specific
releases). The CI/test env still operates as a canary of upstream
changes, but on a delay dependent on upstream releases compared to
current state.