Fix for create_network in interface API, properly deserializing AlchemicalNetwork
Previously, we were relying on `fastapi` to decode the JSON form of the
request, which was not using `gufe.tokenization.JSON_HANDLER.decoder`.
This resulted in `Settings` objects not being turned into `pydantic`
models from `dict` form, which silently worked until now before changes
in `gufe` for 1.0, in which `Protocol` now requires `settings` to be an
actual `Settings` object on init.