Updated cache parameter handling in AlchemiscaleBaseClient
The AlchemiscaleBaseClient now determines the cache directory
when one is not specified directly (i.e. a None is provided to the
AlchemiscaleBaseClient constructor). When a path to this directory is
provided, it must be a string or pathlib.Path object. The logic for
this operation lies in the `AlchemiscaleBaseClient._determine_cache_dir`
method, which can raise a TypeError on invalid input.

The `cache_size_limit` is now verified within the constructor to be
>= 0. If it is not, then a ValueError is raised.

New tests have been added for the above changes:

* Negative cache_size_limit: checks for constructor-raised ValueError
  with a meaningful message.

* cache_directory is None: checks output of the underlying
  _determine_cache_dir method with and without the XDG_CACHE_HOME
  environment variable. If we test it with the client constructor, the
  directory is made automatically, which we don't want in the tests as
  it may touch real data.

* cache_directory is not None, str, or Path: Check that the constructor
  raises a TypeError with a meaningful message.