Fixed the query_network method in `statestore.py`
Since attributes of GufeTokenizables that are set to None are not
explicitly stored in the Neo4j database, regular expression fails to
match nodes. Because of this, we no longer introduce wildcards when a
attribute pattern is set to None. Instead, we completely remove that
attribute regex condition from the WHERE clause.