Fixed bug in n4js_preloaded for compute testing
Since the "incomplete" network's edges are a subset of the full network,
we need to select transformations from the "incomplete" network. Without
this, there is a small chance that a Transformation is not part of the
incomplete network, which, in our tests we always assume the attached
TaskHub has three available tasks available to claim. Additionally, I've
added a couple more assertions checking these assumptions.