Adopt code from py2neo to remove the use of the `id` function in cypher queries (#302)
We are adopting the query generators from within py2neo and modifying them 
to satisfy our needs in alchemiscale. This is needed because the `id` function 
is deprecated, flooding outputs with deprecation warnings.

Code included from py2neo is under the Apache v2.0 license and therefore
we need to include a copy of Apache v2.0 along with a NOTICE with
information on our changes to the code.