Remove throwaway assingment from tests
The original change was intended to discourage developers from removing
an "unused" fixture from test arguments. The fixture itself is not pure
in a sense that it produces artifacts. This change splits the original
fixture into a pure fixture that returns the client and must be used,
and a fixture with a private name that sets the environment variables. I
think it is more idiomatic this way