Adjust get_transformation_chemicalsystems tests
The tests for AlchemiscaleClient.get_transformation_chemicalsystems
and Neo4jStore.get_transformation_chemicalsystems previously checked
the lengths of the returned list of ChemicalSystem ScopedKeys and
compared their contents with a set of the expected ScopedKeys. Since
we now return a deliberate ordering, we can remove the size check and
compare to a list.