Remove DISTINCT keyword from get_keyed_chain query
Originally, the idea was that if we were to ever have a case where a tokenizable had two fields pointing to the same dependency this would remove the redundant processing. However, with a second look over, I think this introduces a bug since later on I used zip with the rels and keys. This means that if any keys are trimmed then I'm dropping relationship information from the resulting keyed_dict.