Returned elements of `get_transformation_chemicalsystems` ordered by `ChemicalSystem` state (#402)
* Order ChemicalSystems keys by their Transformation attribute name

Previously, the order of the returned keys was random. Now we sort by
the attribute name (i.e. "stateA" and "stateB").

* Adjust get_transformation_chemicalsystems tests

The tests for AlchemiscaleClient.get_transformation_chemicalsystems
and Neo4jStore.get_transformation_chemicalsystems previously checked
the lengths of the returned list of ChemicalSystem ScopedKeys and
compared their contents with a set of the expected ScopedKeys. Since
we now return a deliberate ordering, we can remove the size check and
compare to a list.

* Added new entry

* Updated docstring for AlchemiscaleClient.get_transformation_chemicalsystems

* Update tests to check both the Transformation and NonTransformation case

* Update docstring format

* String comparison fix in tests

* Improve docstring clarity for get_transformation_chemicalsystems

Co-authored-by: David L. Dotson <dotsdl@gmail.com>

---------

Co-authored-by: David L. Dotson <dotsdl@gmail.com>