Drop neo4j driver notifications completely
Our use of neo4j results in many notifications of level WARNING coming
back from the server, usually because there are properties are labels in
our queries that do not (yet) have instances in the database. This
results in a flood of warnings that appear to drive up system load on
the host.